More JEST

+++ packages/api/routes/commentaires.js
@@ -1,14 +1,16 @@
-var express = require("express");
-var router = express.Router();
+const express = require("express");
+const createError = require("http-errors");
+const router = express.Router();
 
 const { typeRequired } = require("../auth/_helpers");
 
 const {
   getAllCommentaires,
   addCommentaire,
-  deleteCommentaire,
-  isMandataireInTi
+  deleteCommentaire
 } = require("../db/queries/commentaires");
+
+const { isMandataireInTi } = require("../db/queries/mandataires");
 
 const { getTiByUserId } = require("../db/queries/tis");
 
@@ -34,17 +36,23 @@ router.get(
   "/:mandataireId/commentaires",
   typeRequired("ti"),
   async (req, res, next) => {
-    const ti = await getTiByUserId(req.user.id);
-    if (!ti) {
-      next(new Error("500"));
+    try {
+      const ti = await getTiByUserId(req.user.id);
+      if (!ti) {
+        throw createError.Unauthorized(`TI not found`);
+      }
+      const isAllowed = await isMandataireInTi(req.params.mandataireId, ti.id);
+      if (!isAllowed) {
+        throw createError.Unauthorized(`Mandataire not found in current TI`);
+      }
+      const commentaires = await getAllCommentaires(
+        req.params.mandataireId,
+        ti.id
+      );
+      res.status(200).json(commentaires);
+    } catch (err) {
+      next(err);
     }
-    getAllCommentaires(req.params.mandataireId, ti.id)
-      .then(function(commentaires) {
-        res.status(200).json(commentaires);
-      })
-      .catch(function(error) {
-        next(error);
-      });
   }
 );
 
@@ -87,35 +95,30 @@ router.get(
 router.post(
   "/:mandataireId/commentaires",
   typeRequired("ti"),
-  (req, res, next) => {
+  async (req, res, next) => {
     // ensure TI can write on this mandataire + add related test
-    getTiByUserId(req.user.id).then(ti => {
+    try {
+      const ti = await getTiByUserId(req.user.id);
       if (!ti) {
-        next(new Error("500"));
-        return;
+        throw createError.Unauthorized(`TI not found`);
       }
-      return isMandataireInTi(req.params.mandataireId, ti.id)
-        .then(isMandataireInTi => {
-          if (!isMandataireInTi) {
-            return next(new Error(401));
-          }
-          return addCommentaire({
-            comment: req.body.comment,
-            mandataire_id: req.params.mandataireId,
-            ti_id: ti.id
-          })
-            .then(function() {
-              return getAllCommentaires(req.params.mandataireId, ti.id);
-            })
-            .then(function(commentaires) {
-              res.status(200).json(commentaires);
-            });
-        })
-        .catch(function(error) {
-          console.log(error);
-          next(error);
+      const isAllowed = await isMandataireInTi(req.params.mandataireId, ti.id);
+      if (!isAllowed) {
+        throw createError.Unauthorized(`Mandataire not found in current TI`);
+      }
+      // skip when empty call
+      if (req.body.comment) {
+        await addCommentaire({
+          comment: req.body.comment,
+          mandataire_id: req.params.mandataireId,
+          ti_id: ti.id
         });
-    });
+      }
+      const comments = await getAllCommentaires(req.params.mandataireId, ti.id);
+      res.status(200).json(comments);
+    } catch (err) {
+      next(err);
+    }
   }
 );
 
@@ -155,20 +158,27 @@ router.delete(
   typeRequired("ti"),
   async (req, res, next) => {
     // secu : ensure TI can write on this mandataire + add related test
-    const ti = await getTiByUserId(req.user.id);
-    deleteCommentaire({
-      id: req.params.commentaireId,
-      ti_id: ti.id
-    })
-      .then(function(commentaireID) {
-        return getAllCommentaires(req.params.mandataireId, ti.id);
-      })
-      .then(function(commentaires) {
-        res.status(200).json(commentaires);
-      })
-      .catch(function(error) {
-        next(error);
+    try {
+      const ti = await getTiByUserId(req.user.id);
+      if (!ti) {
+        throw createError.Unauthorized(`TI not found`);
+      }
+      const isAllowed = await isMandataireInTi(req.params.mandataireId, ti.id);
+      if (!isAllowed) {
+        throw createError.Unauthorized(`Mandataire not found in current TI`);
+      }
+      await deleteCommentaire({
+        id: req.params.commentaireId,
+        ti_id: ti.id
       });
+      const commentaires = await getAllCommentaires(
+        req.params.mandataireId,
+        ti.id
+      );
+      res.status(200).json(commentaires);
+    } catch (err) {
+      next(err);
+    }
   }
 );
